requirements.txt
fastapi[standard]>=0.110.0
gspread>=6.0.0
oauth2client>=4.1.3
groq>=0.4.2
pydantic>=2.0.0
python-dotenv>=1.0.1
